Add a knife slider bar, holding the rods that spans the wire.
The rods are simple threaded rod. The bar will be laser-cut from
plywood (15mm: 5x 3mm), so that it is not susceptible to softening
due to warmth over time.

Had to increase the headroom to have enough distance to the idler
wheels, enough slider engagement to not rock back and forth and
enough distance between wire and handle-bar to have enough temperature
gradient that things are cold up there.